refactor: callback queue
