+++ multidb_core.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/redis/go-redis/v9/internal"
+	cbq "github.com/redis/go-redis/v9/internal/callbackqueue"
 	"github.com/redis/go-redis/v9/internal/failuredetector"
 	"github.com/redis/go-redis/v9/internal/otel"
 	"github.com/redis/go-redis/v9/internal/pool"
@@ -43,58 +44,12 @@ type multidbDatabase struct {
 	// firing callbacks for it.
 	removed atomic.Bool
 
-	// cbq delivers user circuit-state callbacks asynchronously (FIFO).
-	cbq callbackQueue
-}
-
-// callbackQueue runs user callbacks on their own goroutine, in FIFO order.
-// Breaker state changes fire deep inside locked sections (failoverMu is held
-// during manual selection, pre-failover probes and AddDatabase probes);
-// invoking a user callback there would self-deadlock the moment the callback
-// touches a control API that takes the same lock.
-type callbackQueue struct {
-	mu       sync.Mutex
-	queue    []func()
-	draining bool
-}
-
-func (q *callbackQueue) dispatch(fn func()) {
-	q.mu.Lock()
-	q.queue = append(q.queue, fn)
-	if q.draining {
-		q.mu.Unlock()
-		return
-	}
-	q.draining = true
-	q.mu.Unlock()
-	go q.drain()
-}
-
-func (q *callbackQueue) drain() {
-	for {
-		q.mu.Lock()
-		if len(q.queue) == 0 {
-			q.draining = false
-			q.mu.Unlock()
-			return
-		}
-		fn := q.queue[0]
-		q.queue = q.queue[1:]
-		q.mu.Unlock()
-		runCallbackSafely(fn)
-	}
-}
-
-// runCallbackSafely keeps a panicking user callback from crashing the process
-// (callbacks run on a library-owned goroutine) and from wedging the queue in
-// the draining state.
-func runCallbackSafely(fn func()) {
-	defer func() {
-		if r := recover(); r != nil {
-			internal.Logger.Printf(context.Background(), "multidb: circuit state callback panicked: %v", r)
-		}
-	}()
-	fn()
+	// cbq delivers user circuit-state callbacks asynchronously (FIFO). Breaker
+	// state changes fire deep inside locked sections (failoverMu is held
+	// during manual selection, pre-failover probes and AddDatabase probes);
+	// invoking a user callback there would self-deadlock the moment the
+	// callback touches a control API that takes the same lock.
+	cbq cbq.CallbackQueue
 }
 
 func (db *multidbDatabase) process(ctx context.Context, cmd Cmder) error {
@@ -240,7 +195,7 @@ type multidbCore struct {
 	// which runs on the background loop; running them inline there would
 	// deadlock if a callback called Close() (close() waits on that goroutine
 	// via wg.Wait). Like the per-db cbq, close() does not wait on this queue.
-	announceQ callbackQueue
+	announceQ cbq.CallbackQueue
 }
 
 func newMultidbCore(opts *MultiDBOptions) *multidbCore {
@@ -330,7 +285,7 @@ func (c *multidbCore) buildDatabase(cfg *MultiDBClientConfig) (*multidbDatabase,
 			// member identity (fqdn) in the callback is the follow-up API
 			// that closes it entirely.
 			from, to := oldState.String(), newState.String()
-			dbRef.cbq.dispatch(func() {
+			dbRef.cbq.Dispatch(func() {
 				if dbRef.removed.Load() {
 					return
 				}
@@ -877,7 +832,7 @@ func (c *multidbCore) switchActive(ctx context.Context, from, to int, reason str
 				// Async via announceQ: running a callback inline here would
 				// deadlock if it called Close(). drain() wraps each in
 				// runCallbackSafely; skip if already closed.
-				c.announceQ.dispatch(func() {
+				c.announceQ.Dispatch(func() {
 					if c.closed.Load() {
 						return
 					}
@@ -886,7 +841,7 @@ func (c *multidbCore) switchActive(ctx context.Context, from, to int, reason str
 			}
 		}
 		if c.opts.OnActiveDatabaseChanged != nil {
-			c.announceQ.dispatch(func() {
+			c.announceQ.Dispatch(func() {
 				if c.closed.Load() {
 					return
 				}
